LPS-184314 Extract service and methods from CompanyIndexFactory to CompanyConcurrentReindexManager and use a Reference to CompanyIndexFactory in CompanyConcurrentReindexManager to use its methods
