+++ modules/apps/portal-search-elasticsearch7/portal-search-elasticsearch7-impl/src/main/java/com/liferay/portal/search/elasticsearch7/internal/index/CompanyIndexFactory.java
@@ -14,33 +14,23 @@
 
 package com.liferay.portal.search.elasticsearch7.internal.index;
 
-import com.liferay.osgi.util.service.Snapshot;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.feature.flag.FeatureFlagManagerUtil;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.Company;
-import com.liferay.portal.kernel.model.CompanyConstants;
 import com.liferay.portal.kernel.service.CompanyLocalService;
 import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.search.ccr.CrossClusterReplicationHelper;
 import com.liferay.portal.search.elasticsearch7.internal.configuration.ElasticsearchConfigurationObserver;
 import com.liferay.portal.search.elasticsearch7.internal.configuration.ElasticsearchConfigurationWrapper;
 import com.liferay.portal.search.elasticsearch7.internal.connection.ElasticsearchConnectionManager;
 import com.liferay.portal.search.elasticsearch7.internal.index.util.IndexFactoryCompanyIdRegistryUtil;
 import com.liferay.portal.search.engine.adapter.SearchEngineAdapter;
 import com.liferay.portal.search.engine.adapter.index.UpdateIndexSettingsIndexRequest;
-import com.liferay.portal.search.index.ConcurrentReindexManager;
 import com.liferay.portal.search.index.IndexNameBuilder;
 import com.liferay.portal.search.spi.model.index.contributor.IndexContributor;
 
-import java.text.SimpleDateFormat;
-
-import java.util.Date;
-
-import org.elasticsearch.action.admin.indices.alias.IndicesAliasesRequest;
 import org.elasticsearch.client.IndicesClient;
-import org.elasticsearch.client.RequestOptions;
 import org.elasticsearch.client.RestHighLevelClient;
 
 import org.osgi.framework.BundleContext;
@@ -52,10 +42,9 @@ import org.osgi.service.component.annotations.Reference;
 /**
  * @author Michael C. Han
  */
-@Component(service = {ConcurrentReindexManager.class, IndexFactory.class})
+@Component(service = IndexFactory.class)
 public class CompanyIndexFactory
-	implements ConcurrentReindexManager, ElasticsearchConfigurationObserver,
-			   IndexFactory {
+	implements ElasticsearchConfigurationObserver, IndexFactory {
 
 	@Override
 	public int compareTo(
@@ -74,37 +63,6 @@ public class CompanyIndexFactory
 		}
 
 		_companyIndexFactoryHelper.createIndex(indexName, indicesClient);
-	}
-
-	@Override
-	public void createNextIndex(long companyId) throws Exception {
-		if (!FeatureFlagManagerUtil.isEnabled("LPS-177664") ||
-			(companyId == CompanyConstants.SYSTEM)) {
-
-			return;
-		}
-
-		String baseIndexName = _indexNameBuilder.getIndexName(companyId);
-
-		SimpleDateFormat dateFormat = new SimpleDateFormat("yyyyMMddHHmmss");
-
-		String timeStampSuffix = dateFormat.format(new Date());
-
-		String newIndexName = baseIndexName + "-" + timeStampSuffix;
-
-		RestHighLevelClient restHighLevelClient =
-			_elasticsearchConnectionManager.getRestHighLevelClient();
-
-		if (_companyIndexFactoryHelper.hasIndex(
-				restHighLevelClient.indices(), newIndexName)) {
-
-			return;
-		}
-
-		_companyIndexFactoryHelper.createIndex(
-			newIndexName, restHighLevelClient.indices());
-
-		_companyLocalService.updateIndexNameNext(companyId, newIndexName);
 	}
 
 	@Override
@@ -132,29 +90,6 @@ public class CompanyIndexFactory
 	}
 
 	@Override
-	public void deleteNextIndex(long companyId) {
-		if (!FeatureFlagManagerUtil.isEnabled("LPS-177664")) {
-			return;
-		}
-
-		Company company = _companyLocalService.fetchCompany(companyId);
-
-		if (company == null) {
-			return;
-		}
-
-		String indexName = company.getIndexNameNext();
-
-		if (!Validator.isBlank(indexName)) {
-			RestHighLevelClient restHighLevelClient =
-				_elasticsearchConnectionManager.getRestHighLevelClient();
-
-			_companyIndexFactoryHelper.deleteIndex(
-				indexName, restHighLevelClient.indices(), companyId, false);
-		}
-	}
-
-	@Override
 	public int getPriority() {
 		return 3;
 	}
@@ -169,69 +104,6 @@ public class CompanyIndexFactory
 	@Override
 	public synchronized void registerCompanyId(long companyId) {
 		IndexFactoryCompanyIdRegistryUtil.registerCompanyId(companyId);
-	}
-
-	@Override
-	public void replaceCurrentIndexWithNextIndex(long companyId)
-		throws Exception {
-
-		if (!FeatureFlagManagerUtil.isEnabled("LPS-177664") ||
-			(companyId == CompanyConstants.SYSTEM)) {
-
-			return;
-		}
-
-		RestHighLevelClient restHighLevelClient =
-			_elasticsearchConnectionManager.getRestHighLevelClient();
-
-		IndicesAliasesRequest indicesAliasesRequest =
-			new IndicesAliasesRequest();
-
-		IndicesAliasesRequest.AliasActions addAliasActions =
-			IndicesAliasesRequest.AliasActions.add();
-
-		String baseIndexName = _indexNameBuilder.getIndexName(companyId);
-
-		addAliasActions.alias(baseIndexName);
-
-		Company company = _companyLocalService.getCompany(companyId);
-
-		String indexNameNext = company.getIndexNameNext();
-
-		addAliasActions.index(indexNameNext);
-
-		indicesAliasesRequest.addAliasAction(addAliasActions);
-
-		String removeIndex = baseIndexName;
-
-		if (!Validator.isBlank(company.getIndexNameCurrent())) {
-			removeIndex = company.getIndexNameCurrent();
-		}
-
-		IndicesAliasesRequest.AliasActions removeIndexAliasActions =
-			IndicesAliasesRequest.AliasActions.removeIndex();
-
-		removeIndexAliasActions.index(removeIndex);
-
-		indicesAliasesRequest.addAliasAction(removeIndexAliasActions);
-
-		IndicesClient indicesClient = restHighLevelClient.indices();
-
-		CrossClusterReplicationHelper crossClusterReplicationHelper =
-			_crossClusterReplicationHelperSnapshot.get();
-
-		if (crossClusterReplicationHelper != null) {
-			crossClusterReplicationHelper.unfollow(removeIndex);
-		}
-
-		indicesClient.updateAliases(
-			indicesAliasesRequest, RequestOptions.DEFAULT);
-
-		_companyLocalService.updateIndexNames(companyId, indexNameNext, null);
-
-		if (crossClusterReplicationHelper != null) {
-			crossClusterReplicationHelper.follow(indexNameNext);
-		}
 	}
 
 	@Override
@@ -322,11 +194,6 @@ public class CompanyIndexFactory
 
 	private static final Log _log = LogFactoryUtil.getLog(
 		CompanyIndexFactory.class);
-
-	private static final Snapshot<CrossClusterReplicationHelper>
-		_crossClusterReplicationHelperSnapshot = new Snapshot(
-			CompanyIndexFactory.class, CrossClusterReplicationHelper.class,
-			null, true);
 
 	@Reference
 	private CompanyIndexFactoryHelper _companyIndexFactoryHelper;
